Add generic subscriber recieve filtering
Sometimes it's possible to save some work by deciding before the message
is decoded to not decode it. Do this having a callback function that
only takes generic metadata about the message.

The use case this is specifically being added for is when there are
multiple UAVCAN node classes listening to different sources of data and
the application needs to chose which node it is actually going to listen
to.