Progress on #251, pool allocator. (#262)
This change reintroduces the pool allocator defined in v0 but designed
in a way that allows sharing pool memory with stl containers. We also
provide a lock-free implementation at the cost of requiring atomics
support. We'll have to come back to add support for cores without CAS
instructions (e.g. Cortex M0).