mass upstream (#326)
Sync Amazon internal working copy with upstream.

Note, this is still very rough and will completely change. Do not use.

I'm forcing this through since a detailed review to a garage submission for a half-baked commit would be a waste. 

We all know we have a lot of work to do here including basic design of the APIs.