progress on Transport layer
Bulk change over from previous design to pycyphal-like design. At this checkin we confirmed the network layer was functional on osx. There should be little difference, if any, for linux. The transport layer is 90% complete with the read and write runFor methods being the last integration point needed for an end-to-end test.