First draft of Registry (#388)
Please note that target branch is not `main` but `issue/registry` - this
first PR is just initial commit to the whole "registry" topic/feature -
not yet workable. My idea is keep feature under `issue/registry` until
it will be ready functional wise, and then post it for Amazon review
when it start working (including examples).