Repair the build (#442)
This is a response to
https://forum.opencyphal.org/t/building-libcyphal/2427

- Added brief build instructions to the README. The point of this change
is mostly to indicate that one does not need to "build" libcyphal to use
it because it is header-only. Building is only needed for development or
verification.
- Fixed broken coverage measurement and updated CI workflow
dependencies. Oh boy.
- Fixed the missing cavl error by pulling cavl from its own repository.
Libcanard does no longer ship cavl and libudpard will eventually be
updated to not do that either.
- The libcanard reference was updated to `master` instead of `v4`,
temporarily until v4 is out. The code that leverages libcanard has been
doctored to accommodate API changes in v4 (still officially unstable but
I am not planning on changing it so it's kinda stable after all).
- Treat `NO_STATIC_ANALYSIS` properly.